telegram_quiz/requirements.txt
python-telegram-bot>=21.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
